Move uid services to their own file
